+++ addons/bus/models/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from . import bus
 from . import bus_presence
+from . import ir_model
 from . import res_users
 from . import res_partner
