+++ addons/bus/static/tests/helpers/mock_server.js
@@ -1,5 +1,8 @@
 /** @odoo-module **/
 
+import { TEST_USER_IDS } from "@bus/../tests/helpers/test_constants";
+import { getPyEnv } from "@bus/../tests/helpers/mock_python_environment";
+
 import { patch } from "@web/core/utils/patch";
 import { MockServer } from "@web/../tests/helpers/mock_server";
 import { makeDeferred } from "@web/../tests/helpers/utils";
@@ -7,8 +10,19 @@ import { makeDeferred } from "@web/../tests/helpers/utils";
 patch(MockServer.prototype, 'bus', {
     init() {
         this._super(...arguments);
+        Object.assign(this, TEST_USER_IDS);
         this.pendingLongpollingPromise = null;
         this.notificationsToBeResolved = [];
+    },
+
+    /**
+     * @override
+     */
+    async setup() {
+        this.pyEnv = await getPyEnv();
+        // link the pyEnv to the actual mockServer after execution of
+        // createWebClient.
+        this.pyEnv.mockServer = this;
     },
 
     //--------------------------------------------------------------------------
