+++ addons/bus/static/tests/helpers/test_constants.js
@@ -1,0 +1,9 @@
+/** @odoo-module **/
+
+export const TEST_USER_IDS = {
+    partnerRootId: 2,
+    currentPartnerId: 3,
+    currentUserId: 2,
+    publicPartnerId: 4,
+    publicUserId: 3,
+};
