+++ addons/hr_holidays/static/tests/helpers/model_definitions_setup.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
 
-import { insertModelFields } from '@mail/../tests/helpers/model_definitions_helpers';
+import { insertModelFields } from '@bus/../tests/helpers/model_definitions_helpers';
 
 insertModelFields('res.partner', {
     out_of_office_date_end: { type: 'date' },
