+++ addons/mail/controllers/discuss.py
@@ -203,12 +203,6 @@ class DiscussController(http.Controller):
     def mail_load_message_failures(self, **kwargs):
         return request.env.user.partner_id._message_fetch_failed()
 
-    @http.route('/mail/get_model_definitions', methods=['POST'], type='http', auth='user')
-    def get_model_definitions(self, model_names_to_fetch, **kwargs):
-        return request.make_response(json.dumps(
-            request.env['ir.model']._get_model_definitions(json.loads(model_names_to_fetch)),
-        ))
-
     # --------------------------------------------------------------------------
     # Mailbox
     # --------------------------------------------------------------------------
