+++ addons/mail/static/tests/helpers/test_utils.js
@@ -1,23 +1,18 @@
 /** @odoo-module **/
 
+import { getPyEnv, startServer } from '@bus/../tests/helpers/mock_python_environment';
+
 import { nextTick } from '@mail/utils/utils';
 import { getAdvanceTime } from '@mail/../tests/helpers/time_control';
 import { getWebClientReady } from '@mail/../tests/helpers/webclient_setup';
 
-import { registry } from '@web/core/registry';
 import { wowlServicesSymbol } from "@web/legacy/utils";
 import { registerCleanup } from "@web/../tests/helpers/cleanup";
-import { MockServer } from "@web/../tests/helpers/mock_server";
 import { getFixture, makeDeferred, patchWithCleanup } from "@web/../tests/helpers/utils";
 import { doAction, getActionManagerServerData } from "@web/../tests/webclient/helpers";
 
-import core from 'web.core';
-
 const { App, EventBus } = owl;
 const { afterNextRender } = App;
-const modelDefinitionsPromise = new Promise(resolve => {
-    QUnit.begin(() => resolve(getModelDefinitions()));
-});
 
 //------------------------------------------------------------------------------
 // Private
@@ -53,215 +48,6 @@ function nextAnimationFrame() {
     return new Promise(function (resolve) {
         setTimeout(() => requestAnimationFrame(() => resolve()));
     });
-}
-
-//------------------------------------------------------------------------------
-// Model definitions setup
-//------------------------------------------------------------------------------
-
-export const TEST_USER_IDS = {
-    partnerRootId: 2,
-    currentPartnerId: 3,
-    currentUserId: 2,
-    publicPartnerId: 4,
-    publicUserId: 3,
-};
-
-/**
- * Fetch model definitions from the server then insert fields present in the
- * `mail.model.definitions` registry. Use `addModelNamesToFetch`/`insertModelFields`
- * helpers in order to add models to be fetched, default values to the fields,
- * fields to a model definition.
- *
- * @return {Map<string, Object>} A map from model names to model fields definitions.
- * @see model_definitions_setup.js
- */
-async function getModelDefinitions() {
-    const modelDefinitionsRegistry = registry.category('mail.model.definitions');
-    const modelNamesToFetch = modelDefinitionsRegistry.get('modelNamesToFetch');
-    const fieldsToInsertRegistry = modelDefinitionsRegistry.category('fieldsToInsert');
-
-    // fetch the model definitions.
-    const formData = new FormData();
-    formData.append('csrf_token', core.csrf_token);
-    formData.append('model_names_to_fetch', JSON.stringify(modelNamesToFetch));
-    const response = await window.fetch('/mail/get_model_definitions', { body: formData, method: 'POST' });
-    if (response.status !== 200) {
-        throw new Error('Error while fetching required models');
-    }
-    const modelDefinitions = new Map(Object.entries(await response.json()));
-
-    for (const [modelName, fields] of modelDefinitions) {
-         // insert fields present in the fieldsToInsert registry : if the field
-         // exists, update its default value according to the one in the
-         // registry; If it does not exist, add it to the model definition.
-        const fieldNamesToFieldToInsert = fieldsToInsertRegistry.category(modelName).getEntries();
-        for (const [fname, fieldToInsert] of fieldNamesToFieldToInsert) {
-            if (fname in fields) {
-                fields[fname].default = fieldToInsert.default;
-            } else {
-                fields[fname] = fieldToInsert;
-            }
-        }
-        // apply default values for date like fields if none was passed.
-        for (const fname in fields) {
-            const field = fields[fname];
-            if (['date', 'datetime'].includes(field.type) && !field.default) {
-                const defaultFieldValue = field.type === 'date'
-                    ? () => moment.utc().format('YYYY-MM-DD')
-                    : () => moment.utc().format("YYYY-MM-DD HH:mm:ss");
-                field.default = defaultFieldValue;
-            } else if (fname === 'active' && !('default' in field)) {
-                // records are active by default.
-                field.default = true;
-            }
-        }
-    }
-    // add models present in the fake models registry to the model definitions.
-    const fakeModels = modelDefinitionsRegistry.category('fakeModels').getEntries();
-    for (const [modelName, fields] of fakeModels) {
-        modelDefinitions.set(modelName, fields);
-    }
-    return modelDefinitions;
-}
-
-let pyEnv;
-/**
- * Creates an environment that can be used to setup test data as well as
- * creating data after test start.
- *
- * @returns {Object} An environment that can be used to interact with the mock
- * server (creation, deletion, update of records...)
- */
- export async function startServer() {
-    const models = {};
-    const views = {};
-    const modelDefinitions = await modelDefinitionsPromise;
-    const recordsToInsertRegistry = registry.category('mail.model.definitions').category('recordsToInsert');
-    for (const [modelName, fields] of modelDefinitions) {
-        const records = [];
-        if (recordsToInsertRegistry.contains(modelName)) {
-            // prevent tests from mutating the records.
-            records.push(...JSON.parse(JSON.stringify(recordsToInsertRegistry.get(modelName))));
-        }
-        models[modelName] = { fields: { ...fields }, records };
-
-        // generate default views for this model.
-        const viewArchsSubRegistries = registry.category('mail.view.archs').subRegistries;
-        for (const [viewType, archsRegistry] of Object.entries(viewArchsSubRegistries)) {
-            views[`${modelName},false,${viewType}`] =
-                archsRegistry.get(modelName, archsRegistry.get('default'));
-        }
-    }
-    pyEnv = new Proxy(
-        {
-            getData() {
-                return this.mockServer.models;
-            },
-            getViews() {
-                return views;
-            },
-            ...TEST_USER_IDS,
-            get currentPartner() {
-                return this.mockServer.currentPartner;
-            },
-        },
-        {
-            get(target, name) {
-                if (target[name]) {
-                    return target[name];
-                }
-                const modelAPI = {
-                    /**
-                     * Simulate a 'create' operation on a model.
-                     *
-                     * @param {Object[]|Object} values records to be created.
-                     * @returns {integer[]|integer} array of ids if more than one value was passed,
-                     * id of created record otherwise.
-                     */
-                    create(values) {
-                        if (!values) {
-                            return;
-                        }
-                        if (!Array.isArray(values)) {
-                            values = [values];
-                        }
-                        const recordIds = values.map(value => target.mockServer.mockCreate(name, value));
-                        return recordIds.length === 1 ? recordIds[0] : recordIds;
-                    },
-                    /**
-                     * Simulate a 'search' operation on a model.
-                     *
-                     * @param {Array} domain
-                     * @param {Object} context
-                     * @returns {integer[]} array of ids corresponding to the given domain.
-                     */
-                    search(domain, context = {}) {
-                        return target.mockServer.mockSearch(name, [domain], context);
-                    },
-                    /**
-                     * Simulate a `search_count` operation on a model.
-                     *
-                     * @param {Array} domain
-                     * @return {number} count of records matching the given domain.
-                     */
-                    searchCount(domain) {
-                        return this.search(domain).length;
-                    },
-                    /**
-                     * Simulate a 'search_read' operation on a model.
-                     *
-                     * @param {Array} domain
-                     * @param {Object} kwargs
-                     * @returns {Object[]} array of records corresponding to the given domain.
-                     */
-                    searchRead(domain, kwargs = {}) {
-                        return target.mockServer.mockSearchRead(name, [domain], kwargs);
-                    },
-                    /**
-                     * Simulate an 'unlink' operation on a model.
-                     *
-                     * @param {integer[]} ids
-                     * @returns {boolean} mockServer 'unlink' method always returns true.
-                     */
-                    unlink(ids) {
-                        return target.mockServer.mockUnlink(name, [ids]);
-                    },
-                    /**
-                     * Simulate a 'write' operation on a model.
-                     *
-                     * @param {integer[]} ids ids of records to write on.
-                     * @param {Object} values values to write on the records matching given ids.
-                     * @returns {boolean} mockServer 'write' method always returns true.
-                     */
-                    write(ids, values) {
-                        return target.mockServer.mockWrite(name, [ids, values]);
-                    },
-                };
-                if (name === 'bus.bus') {
-                    modelAPI['_sendone'] = target.mockServer._mockBusBus__sendone.bind(target.mockServer);
-                    modelAPI['_sendmany'] = target.mockServer._mockBusBus__sendmany.bind(target.mockServer);
-                }
-                return modelAPI;
-            },
-            set(target, name, value) {
-                return target[name] = value;
-            },
-         },
-    );
-    pyEnv['mockServer'] = new MockServer({ models }, {});
-    await pyEnv['mockServer'].setup();
-    registerCleanup(() => pyEnv = undefined);
-    return pyEnv;
-}
-
-/**
- *
- * @returns {Object} An environment that can be used to interact with the mock
- * server (creation, deletion, update of records...)
- */
-export function getPyEnv() {
-    return pyEnv || startServer();
 }
 
 //------------------------------------------------------------------------------
@@ -460,7 +246,7 @@ async function start(param0 = {}) {
         waitUntilEventPromise = Promise.resolve();
     }
 
-    pyEnv = await getPyEnv();
+    const pyEnv = await getPyEnv();
     param0.serverData = param0.serverData || getActionManagerServerData();
     param0.serverData.models = { ...pyEnv.getData(), ...param0.serverData.models };
     param0.serverData.views = { ...pyEnv.getViews(), ...param0.serverData.views };
@@ -482,8 +268,6 @@ async function start(param0 = {}) {
         await webClient.env.services.messaging.modelManager.messagingCreatedPromise;
         await webClient.env.services.messaging.modelManager.messagingInitializedPromise;
     }
-    // link the pyEnv to the actual mockServer after execution of createWebClient.
-    pyEnv.mockServer = MockServer.currentMockServer;
     const openView = async (action, options) => {
         action['type'] = action['type'] || 'ir.actions.act_window';
         await afterNextRender(() => doAction(webClient, action, { props: options }));
@@ -603,4 +387,5 @@ export {
     nextTick,
     pasteFiles,
     start,
+    startServer,
 };
