+++ addons/snailmail/static/tests/helpers/model_definitions_setup.js
@@ -1,5 +1,5 @@
 /** @odoo-module **/
 
-import { addModelNamesToFetch } from '@mail/../tests/helpers/model_definitions_helpers';
+import { addModelNamesToFetch } from '@bus/../tests/helpers/model_definitions_helpers';
 
 addModelNamesToFetch(['snailmail.letter']);
