+++ addons/website_livechat/static/tests/helpers/model_definitions_setup.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
 
-import { addModelNamesToFetch, insertModelFields } from '@mail/../tests/helpers/model_definitions_helpers';
+import { addModelNamesToFetch, insertModelFields } from '@bus/../tests/helpers/model_definitions_helpers';
 
 addModelNamesToFetch(['website.visitor']);
 insertModelFields('mail.channel', {
