+++ addons/website_slides/static/tests/helpers/view_definitions_setup.js
@@ -2,7 +2,7 @@
 
 import { registry } from '@web/core/registry';
 
-const viewArchsRegistry = registry.category('mail.view.archs');
+const viewArchsRegistry = registry.category('bus.view.archs');
 
 viewArchsRegistry.category('form').add(
     'slide.channel',
